debug formulaire modif infouser

+++ w/public/assets/js/edit-video.js
@@ -2,23 +2,27 @@
 
 $(function () {
 
-    $(document).on('click','.deleteVideo',function () {
+    $('[data-toggle=confirmation]').confirmation({
+        rootSelector: '[data-toggle=confirmation]',
 
-        var id = $(this).data('delete');
-        $.ajax({
-            type:'POST',
-            url:$('.deleteId').val(),
-            data:{
-                id: id
-            }
-        }).done(function () {
-            deleteElement(id);
-            if($(".latest-video").children().length === 0){
-                window.location.reload();
-            }
-        }).fail(function (response) {
-            console.log(response);
-        })
+        onConfirm: function() {
+            var id = $(this).data('delete');
+            $.ajax({
+                type:'POST',
+                url:$('.deleteId').val(),
+                data:{
+                    id: id
+                }
+            }).done(function () {
+                deleteElement(id);
+                if($(".latest-video").children().length === 0){
+                    window.location.reload();
+                }
+            }).fail(function (response) {
+                console.log(response);
+            })
+        },
+        // other options
     });
 
     $(document).on('click', '.editVideo',function () {
